# api-gateway/docs/swagger.yaml
definitions:
...
  genprotos.JobApplicationCreate:
    properties:
      candidate_id:
        type: string
      resume_id:
        type: string
      status:
        type: string
      vacancy_id:
        type: string
    type: object
  genprotos.JobApplicationGetAllRes:
    properties:
      JobApplications:
        $ref: '#/definitions/genprotos.JobApplitcationGetRes'
      success:
        type: boolean
    type: object
...
  genprotos.JobApplicationUpdate:
    properties:
      id:
        type: string
      resume_id:
        type: string
      status:
        type: string
    type: object
  genprotos.JobApplitcationGetRes:
    properties:
      candidate_id:
        type: string
      id:
        type: string
      resume_id:
        type: string
      status:
        type: string
      vacancy_id:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /job-applications:
    post:
      consumes:
      - application/json
      description: Create a new job application
      parameters:
      - description: Job Application data
        in: body
        name: job_application
        required: true
        schema:
          $ref: '#/definitions/genprotos.JobApplicationCreate'
      produces:
      - application/json
      responses:
        "200":
          description: Job application created
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create job application
      tags:
      - jobapplication
  /job-applications/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a job application by ID
      parameters:
      - description: Job Application ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Job application deleted
          schema:
            type: string
        "400":
          description: Invalid job application ID
          schema:
            type: string
        "404":
          description: Job application not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete job application
      tags:
      - jobapplication
    get:
      consumes:
      - application/json
      description: Get a job application by its ID
      parameters:
      - description: Job Application ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genprotos.JobApplitcationGetRes'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get job application by ID
      tags:
      - jobapplication
    put:
      consumes:
      - application/json
      description: Update an existing job application
      parameters:
      - description: Job Application ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated job application data
        in: body
        name: job_application
        required: true
        schema:
          $ref: '#/definitions/genprotos.JobApplicationUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: Job application updated
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "404":
          description: Job application not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update job application
      tags:
      - jobapplication
  /job-applications/all:
    get:
      consumes:
      - application/json
      description: Get all job applications
      parameters:
      - description: Candidate ID
        in: query
        name: candidate_id
        type: string
      - description: Vacancy ID
        in: query
        name: vacancy_id
        type: string
      - description: Resume ID
        in: query
        name: resume_id
        type: string
      - description: Status
        in: query
        name: status
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genprotos.JobApplicationGetAllRes'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all job applications
      tags:
      - jobapplication
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"